Add `status` field to invoices, as per upcoming changes in c-lightning.
Three possible states: `unpaid`, `paid` and `expired`.

The `completed` field is now deprecated as a public field
and will soon be hidden from the public APIs.

However, `completed` will continue to be used for internal storage,
with `status` as a virtual field determined based on `completed`, `expiry_time`
and the current time serving the request.