Don't run autocleanup if disabled on c-lightning
We cleanup old invoices only after checking they don't exists on c-lightning.
If autocleanup is enabled on charge but not on c-lightning, it'll check the
same invoices over and over again, wasting a lot of resources and never actually
deleting them.

This commit makes Charge only enable autocleanup when its enabled on
c-lightning and use the same configurations for the cycle interval
and how long to keep expired invoices for.